docs: update branching documentation to Trunk Based.
There was still some legacy documentation of a method we did not
use at all, and probably will never use. I updated it to the
method closest to what we are using, and updated the links.

Also removed a broken link to an image we did not have, and
decided to not include an Trunk Based image since we have
no images at all in our CONTRIBUTING.

Signed-off-by: Daniël Kok <daniel.kok@alliander.com>