docs: remove incorrect branching structure from contributing (#26)
There was still some legacy documentation of a method we did not
use at all, and probably will never use. I updated it to the
method closest to what we are using, and updated the links.

Also removed a broken link to an image we did not have, and
decided to not include an Trunk Based image since we have
no images at all in our CONTRIBUTING.

Since I was there anyway, I further clarified the ways of
contributing, adding the cool schematic from our presentation,
and added an anchor for the commit naming convention to
easily link to.

---------

Signed-off-by: Daniël Kok <daniel.kok@alliander.com>