docs: add a state diagram showing the workflow of ttm.
I opted for a state diagram instead of a class diagram or
architecture diagram for a couple of reasons. Firstly, the
architecture diagram is quite chaotic from Mermaid, and is
mainly for architecture of web apps and stuff. The class
diagram is mainly focused on class inheritance, and also has
no real flow to it.

My goal of this diagram was for someone to be able to visually
read the diagramand understand the flow of the process, and
what they needed to provide and where these items were used
in the codebase.

Signed-off-by: Daniël Kok <daniel.kok@alliander.com>