docs: implement mermaid architecture into the mkdocs page.
I chose to move the entire .mmd file into a .md file. I also
added a plugin to generate the mermaid diagram, and it
automatically inherits the theme colours of Material. I don't
really fancy that, as the theme colors from Alliander are
ugly, but it'll have to do because I don't really know
how to change this.

I did keep the coloring of the different states because
that makes the diagram way more readable. I also opted for
a LR direction of the Data Collection so that it is better
readable on a vertical page. Otherwise the figure would have
been to far zoomed out and it would've been unreadable.

Signed-off-by: Daniël Kok <daniel.kok@alliander.com>