ci: also publish to github releases on release workflow (#43)
NOTE: I used copilot to help me with generating the changelog. I want
the releases on our github page to coincide with the releases on PyPI. I
expanded the release workflow to build the package again, generate a
changelog based on all the merged pr's since previous release, and then
create the github release.

Since we have no option for pre-releases yet build into this workflow, I
opted for hard-coding the pre-release option set to 'false'.

The rest of the format for this release note is how I know it from other
open-source projects.

Signed-off-by: Daniël Kok <daniel.kok@alliander.com>