fix!: split hotspot increase into winding and oil part (#39)
fix #32. Split the calculation of the hotspot into two parts. The
increase due the the windings, and the oil. This fixes a bug where the
formule did not follow the one described by IEC 60076-7 (2018). To check
this, a test is added to compares the results predicted by the IEC and
the results of our model.

BREAKING: As a result of the changes in the formulas, an initial hotspot
temperature can no longer be used in the calculations.

also, a small bug is fixed where "_calculate_internal_temp" suggests a
pandas series, which should be a numpy array.

---------

Signed-off-by: Harm <harm.van.leijen@alliander.com>
Signed-off-by: Daniël Kok <daniel.kok@alliander.com>
Signed-off-by: Daniël Kok <github@danieljkok.nl>
Co-authored-by: Daniël Kok <daniel.kok@alliander.com>