Feature/split input profile (#60)
This is the second pull request to incorporate the three-winding
transformer into the model. In this update, the input profile has been
split into a parent and child class, similar to how the transformer
specifications were previously separated. This change is necessary
because three-winding transformers require three input profile; one each
for high, medium, and low voltage.

Additionally, a new property, "load_profile_array," has been added to
the class. This property provides an array representation of the input
profile: it is 1D for two-winding transformers and 2D for three-winding
transformers. This structure will facilitate future model calculations.

The update is fully backwards compatible. The only modification to the
tests was an example where the return order was adjusted.

---------

Signed-off-by: Harm <harm.van.leijen@alliander.com>