Feaute/create three winding transformer (#61)
This Pr contains two changes:

1) : The naming of three phase has been changed to three winding.

2) The introduction of the three-phase transformer to the package. The
calculation for the end temperature of the top-oil has been adapted
specifically for this transformer type. For now, the default parameters
are the same as those used for power transformers. Note that transformer
component calculations for three-winding transformers have not yet been
implemented, as it is still under discussion whether this functionality
will be included in the model.

The model has also been slightly modified to allow the new vectorised
calculation with the load for three phase transformers.

---------

Signed-off-by: Harm <harm.van.leijen@alliander.com>