Feature/update model for three winding (#62)
Here we change the model class to handle the calculations for the three
winding transformer. to do this we splitted the calculation fo the
top_oil temperature and the hotspot temperature. This was done since the
calculation of the top_oil stayed mostly the same, while the
calculations of the hotspot required an additional for loop, to loop
over the different load_profiles.

---------

Signed-off-by: Harm <harm.van.leijen@alliander.com>