refactor: change if-statement to match-case. (#104)
This improves the interaction with mypy, and allows the ability of
adding the 'assert_never' statement, that should help us catch the
moments where we add something to the enum PaperInsulationType and
forget to include it in this function.

Since so few lines were changed, i NEEDED to add a unit-test for this.
It's a very redundant unit test since were catching everything through typing
already, but oh well, please the system.

Added this after checking this file through mutation tests, which will
have a pr after this one. (#106)

---------

Signed-off-by: Daniël Kok <daniel.kok@alliander.com>